refactor: manage user theme in provider
